Правки по карточке
