+++ js/main.js
@@ -1,6 +1,5 @@
 import {
-  similarAds,
-  newCardData
+  similarAds
 } from './data.js';
 
 
@@ -8,10 +7,9 @@ import {
   createCard
 } from './card.js';
 
-
 /* eslint-disable-next-line no-console */
 console.log(similarAds);
 
 
 const map = document.querySelector('#map-canvas');
-map.appendChild(createCard(newCardData));
+map.appendChild(createCard(similarAds[0]));
